refactor: move XenonDevice and friends to own file
