+++ tinytuya/__init__.py
@@ -102,3 +102,5 @@ from .crypto_helper import *
 from .message_helper import *
 from .exceptions import *
 from .error_helper import *
+from .const import *
+from .XenonDevice import *
